pyproject.toml
[project]
name = "sockio"
version = "0.1.0"
description = "WebSocket chat server"
authors = [
    {name = "devpastor", email = "skadskad13@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core WebSocket and MongoDB dependencies
    "socketify>=0.0.27",
    "beanie>=1.23.0",
    "motor>=3.3.0",
    "pymongo>=4.6.0",
    "pydantic>=2.5.0",
    "pydantic[email]>=2.5.0",
    "watchfiles>=1.1.0",
    # Authentication and security
    "werkzeug>=3.0.0",
    "PyJWT>=2.8.0",
    "bcrypt>=4.1.0",
    # File handling and utilities
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    "Pillow>=10.1.0",
    # HTTP server for REST API endpoints
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    # Configuration and logging
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "pydantic-settings>=2.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sockio"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["sockio"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

